Refactor: extract pto_runtime_c_api shared glue into common (-303 lines)
Move the byte-identical c_api functions from a2a3 and a5 onboard
pto_runtime_c_api.cpp into a single shared
src/common/platform/onboard/host/c_api_shared.cpp linked into each
arch's libhost_runtime.so. Works through DeviceRunnerBase * and
dispatches arch-specific behavior through three new virtuals.

DeviceRunnerBase additions:
- Public virtual destructor (was protected non-virtual). Lets the
  shared destroy_device_context delete polymorphically.
- virtual int run(Runtime&, int, int) = 0  — each arch already had run().
- virtual int finalize() = 0  — each arch already had finalize().
- virtual void set_dep_gen_enabled(bool) — default no-op for a5;
  a2a3 overrides to wire enable_dep_gen_. Lets the shared
  run_prepared call set_dep_gen_enabled unconditionally.

c_api_shared.cpp owns (12 dlsym'd + 7 static helpers + TSD glue):
- TSD glue (g_runner_key, pthread_once, current_runner)
- Static internal: device_malloc, device_free, copy_to_device,
  copy_from_device, upload_chip_callable_buffer_wrapper,
  setup_static_arena_wrapper, acquire_pooled_*_wrapper
- Public C ABI: destroy_device_context, get_runtime_size,
  device_malloc_ctx, device_free_ctx, copy_to_device_ctx,
  copy_from_device_ctx, finalize_device, simpler_init,
  prepare_callable, run_prepared, unregister_callable,
  get_aicpu_dlopen_count, get_host_dlopen_count

Each arch's pto_runtime_c_api.cpp now only carries arch-specific
entries (also dlsym'd by ChipWorker):
- create_device_context (must \`new\` the concrete subclass)
- a2a3: ensure_acl_ready_ctx + create_comm_stream_ctx +
  destroy_comm_stream_ctx (real implementations through
  DeviceRunner::{ensure_acl_ready, create_comm_stream, destroy_comm_stream})
- a5: same three as not-supported stubs plus all comm_* stubs
  (returns NULL / -1, distributed runtime not yet on a5)

Both arches built clean. nm -D verifies all 26 ChipWorker dlsym
targets are exported on both libhost_runtime.so. a2a3 onboard smoke
(dummy_task, alternating_matmul_add, prepared_callable suite,
spmd_basic) — 9/9 passed in 19s.
